backend/requirements.txt
fastapi
pyautogen
